+++ src/com/threewisedroids/belote4droid/base/Trick.java
@@ -10,10 +10,10 @@ public class Trick {
     /**
      * Content of the trick.
      */
-    long[] cards = new long[4];
+    int[] cards = new int[4];
 
     /**
      * Id of the player who started the trick.
      */
-    int    firstPlayer;
+    int   firstPlayer;
 }
